requirements.txt
py_translator==1.8.9
python-docx
tqdm